Added Library Contract to README
One thing that I think is missing from many libraries is the contract
with the end user - that is, in general terms why this library exists
and what the user can expect from it. It doesn't need implementation
details - it should just tell me (or you or whoever) what are reasonable
assumptions that can be made about this library.